refactor(agents-api): Moved some stuff around
